fix: multiline content-type edge case
This fixes a case where the content tiype gets split onto multiple lines in
the verbose window. This causes us to miss the content type and not set the
filetype of the new result window.

This happens when the result gets streamed to stdout in two steps and splits
the header in two, so we just need to check the first line in case we're on
the second step of the stream.